feat: use the new DSL builder for config
