+++ mix.exs
@@ -24,6 +24,7 @@ defmodule AshPostgres.MixProject do
       dialyzer: [
         plt_add_apps: [:ecto]
       ],
+      docs: docs(),
       aliases: aliases(),
       package: package(),
       source_url: "https://github.com/ash-project/ash_postgres",
@@ -41,14 +42,29 @@ defmodule AshPostgres.MixProject do
     ]
   end
 
+  defp docs do
+    [
+      main: "AshPostgres",
+      source_ref: "v#{@version}",
+      logo: "logos/small-logo.png",
+      groups_for_modules: [
+        "entry point": [AshPostgres],
+        "data layer": [AshPostgres.DataLayer],
+        repo: [AshPostgres.Repo],
+        "filter predicates": ~r/AshPostgres.Predicates/
+      ]
+    ]
+  end
+
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
-      {:ecto_sql, "~> 3.0"},
+      {:ecto_sql, "~> 3.4"},
       {:postgrex, ">= 0.0.0"},
-      {:ash, "~> 0.3.0"},
+      {:ash, "~> 0.4.0"},
+      {:db_connection, "~> 2.2", override: true},
       {:git_ops, "~> 2.0.0", only: :dev},
-      {:ex_doc, "~> 0.21", only: :dev, runtime: false},
+      {:ex_doc, "~> 0.22", only: :dev, runtime: false},
       {:ex_check, "~> 0.11.0", only: :dev},
       {:credo, ">= 0.0.0", only: :dev, runtime: false},
       {:dialyxir, ">= 0.0.0", only: :dev, runtime: false},
